requirements.txt
emoji_country_flag==1.3.1
prettytable==3.3.0
pydig==0.4.0
requests==2.28.1
geoip2==4.6.0